pyproject.toml
[tool.poetry]
name = "beer4u-service"
version = "0.1.0"
description = ""
authors = ["Damian Ontivero <damian.ontivero@outlook.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
werkzeug = "^3.0.1"
python-multipart = "^0.0.9"
cryptography = "^42.0.2"
pika = "^1.3.2"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
ipdb = "^0.13.13"
pytest = "^8.0.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
